// src/app/shared/models/modules.model.ts
export class Modules {
  modules_id?: number = 0;
  app_key: string = '';
  name: string = '';
  icon: string = '';
  descs: string = '';
  url: string = '';
  positions: string = '';
  orders: number = 0;
  created_by: string = '';
  created_at: Date = new Date();
  updated_by: string = '';
  updated_at: Date = null;
  flag: number = 1;
}